move function to applier
